refactor plan nodes
