+++ test/unit_tests/backend/plan/run_plan.py
@@ -1,4 +1,5 @@
-from vqpy.backend.plan import Planner, Executor
+from vqpy.backend.executor import Executor
+from vqpy.backend.planner import Planner
 from vqpy.frontend.vobj import VObjBase, vobj_property
 from vqpy.frontend.query import QueryBase
 import os
