--- vqpy/backend/plan/__init__.py
@@ -1,3 +1,0 @@
-from .plan_node import Planner, Executor
-
-__all__ = ["Planner", "Executor"]
